Assignment 4-5/requirements.txt
# Core scientific libraries
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1

# Image processing
pillow==10.4.0
imageio==2.36.0

# Deep learning
tensorflow==2.17.0
keras==3.5.0

# Utility libraries (optional but useful)
matplotlib==3.9.2
tqdm==4.66.5
